+++ gdsfactory/components/mzi.py
@@ -284,7 +284,7 @@ if __name__ == "__main__":
     # c = gf.routing.add_fiber_array(c)
     # gdspath = c.write_gds(flatten_invalid_refs=True)
     # gf.show(gdspath)
-    # c.plot_schematic()
+    # c.plot_schematic_networkx()
     n = c.to_graphviz()
     gf.schematic.plot_graphviz(n)
     c.show()